refactor: extract commandRunnerPlatform helpers into separate module
Moves createCommandRunnerEnv and wrapCommandForPlatform out of commandRunner.js
into a dedicated commandRunnerPlatform.js to improve testability and separation
of concerns. Backward-compatible re-exports are preserved in commandRunner.js.
